Adds Shared-Server integration

+++ src/Response.cpp
@@ -1,22 +1,24 @@
+// "Copyright 2016 <Jobify>"
 #include "Response.h"
 #include <string>
+
 
 Response::Response() {}
 
 std::string Response::getContent() {
-	return content;
+    return content;
 }
 
-long Response::getStatus() {
-	return http_code;
+int Response::getStatus() {
+    return http_code;
 }
 
 void Response::setContent(std::string str) {
-	content = str;
+    content = str;
 }
 
-void Response::setStatus(long code) {
-	http_code = code;
+void Response::setStatus(int code) {
+    http_code = code;
 }
 
 Response::~Response() {}
